fix:add utils file and seperate type evaluation logic in core
